Refactor dev server.
