+++ packages/gensx/src/dev-server/index.ts
@@ -1,0 +1,30 @@
+// Main exports
+export { GensxServer, createServer } from "./server.js";
+
+// Types and interfaces
+export type {
+  ServerOptions,
+  WorkflowInfo,
+  WorkflowExecution,
+  ExecutionStatus,
+  WorkflowMessage,
+  JsonValue,
+  InputRequest,
+} from "./types.js";
+
+// Error classes
+export { NotFoundError, BadRequestError, ServerError } from "./errors.js";
+
+// Utility functions
+export { generateWorkflowId, generateExecutionId } from "./utils.js";
+
+// Manager classes (for advanced usage)
+export { WorkflowManager } from "./workflow-manager.js";
+export { ValidationManager } from "./validation.js";
+export { ExecutionHandler } from "./execution-handler.js";
+
+// OpenAPI utilities
+export { generateOpenApiSpec, generateSwaggerUI } from "./openapi.js";
+
+// Route setup (for advanced usage)
+export { setupRoutes } from "./routes.js";
